requirements.txt
torch>=2.0.0
torchvision>=0.15.0
accelerate>=0.20.0
transformers>=4.40.0
diffusers>=0.27.0
Pillow>=10.0.0
numpy>=1.24.0
PyYAML>=6.0
tqdm>=4.65.0
tensorboard>=2.13.0
wandb>=0.15.0
lpips>=0.1.4
opencv-python>=4.8.0
scikit-image>=0.21.0
pytesseract>=0.3.10
einops>=0.7.0
safetensors>=0.3.1
peft>=0.7.0
bitsandbytes>=0.41.0
